requirements.txt
-e .

## Db
alembic==1.13.1
psycopg2==2.9.9

# Logging
rich==13.3.5
python-json-logger==2.0.7

# Environment
python-dotenv==1.0.0

## Schemas
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3

fastapi==0.110.0
uvicorn==0.30.0
pillow==10.2.0
python-multipart==0.0.9
openai==1.30.0
treys
opencv-python
pytesseract
langchain
langchain-openai
